+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## V0.11.1 (released)
 
-- BREAKING: Usage of createIndex changed `createIndex(uid: string, options: IndexOptions)` #436
+- BREAKING: Usage of createIndex changed `createIndex(uid: string, options: IndexOptions): Index` #436
 - Error Handler improved by adding a new MeiliSearchCommunicationError #436
 - Refactor Error handler #436
 - FEATURE: Add getOrCreateIndex method to meilisearch client #436